Resolves #1240 again
