+++ desc/objectives/_geometry.py
@@ -3,11 +3,18 @@
 import numpy as np
 
 from desc.backend import jnp, vmap
-from desc.compute import get_profiles, get_transforms, rpz2xyz
-from desc.compute.geom_utils import copy_rpz_periods
+from desc.compute import get_profiles, get_transforms
 from desc.compute.utils import _compute as compute_fun
 from desc.grid import LinearGrid, QuadratureGrid
-from desc.utils import Timer, errorif, parse_argname_change, safenorm, warnif
+from desc.utils import (
+    Timer,
+    copy_rpz_periods,
+    errorif,
+    parse_argname_change,
+    rpz2xyz,
+    safenorm,
+    warnif,
+)
 
 from .normalization import compute_scaling_factors
 from .objective_funs import _Objective, collect_docs
