+++ tests/test_magnetic_fields.py
@@ -8,7 +8,6 @@ from scipy.constants import mu_0
 
 from desc.backend import jax, jit, jnp
 from desc.basis import DoubleFourierSeries
-from desc.compute import rpz2xyz, rpz2xyz_vec, xyz2rpz_vec
 from desc.compute.utils import get_params, get_transforms
 from desc.derivatives import FiniteDiffDerivative as Derivative
 from desc.examples import get
@@ -33,7 +32,7 @@ from desc.magnetic_fields import (
 )
 from desc.magnetic_fields._dommaschk import CD_m_k, CN_m_k
 from desc.plotting import poincare_plot
-from desc.utils import dot
+from desc.utils import dot, rpz2xyz, rpz2xyz_vec, xyz2rpz_vec
 
 
 def phi_lm(R, phi, Z, a, m):
